refactor: 템플릿 화면 관련 파일 구조 변경
- 각 템플릿 탭(메인, 내 보따리)에 해당하는 파일을 `main`, `my` 하위 패키지로 이동하여 구조를 명확히 합니다.
- `MainTemplateContent.kt`와 `MyTemplateContent.kt`의 패키지 경로를 변경합니다.
- 북마크 템플릿 화면을 위한 `BookmarkTemplateContent.kt` 파일을 추가합니다.
