+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/compose/home/template/bookmark/BookmarkTemplateContent.kt
@@ -1,0 +1,8 @@
+package com.bottari.presentation.compose.home.template.bookmark
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+@Composable
+fun BookmarkTemplateContent(modifier: Modifier = Modifier) {
+}
